requirements.txt
ipydex
markdown
markdownify >= 0.13
bs4
gitpython
